adding triangles

+++ src/Domain_d.C
@@ -65,7 +65,7 @@ void Domain_d::SetDimension(const int &node_count, const int &elem_count){
 
   malloc_t(m_detJ,  double, m_elem_count * m_gp_count );    
   
-  malloc_t(m_nodxelem_e,  double, m_elem_count);
+  malloc_t(m_nodxelem_e,  int, m_elem_count);
   
   malloc_t(m_ematm, double, m_nodxelem * m_nodxelem * m_elem_count); //Elemental mas matrices
